fixed int error
was converting keys from strings to integers, forgetting that some contain decimals. Fixed by converting them to floats instead. Did have it finished on time but needed to fix this minor error late.

+++ shop.py
@@ -59,7 +59,7 @@ while True:
         if buy in list(BuyDict):    
             for key in list(BuyDict):
                 if key == buy:
-                    cost = int((BuyDict[key]))
+                    cost = float((BuyDict[key]))
                     profit += cost
                     BuyDict.pop(key)
         else:
@@ -70,7 +70,7 @@ while True:
         if sell in list(SellDict):    
             for key in list(SellDict):
                 if key == sell:
-                    cost = int((SellDict[key]))
+                    cost = float((SellDict[key]))
                     profit -= cost
                     SellDict.pop(key)
         else:
